Split off DB testing
